♻️ Move implementations for Open from open.go to factory.go
