+++ open.go
@@ -3,8 +3,6 @@ package sqlslog
 import (
 	"context"
 	"database/sql"
-	"database/sql/driver"
-	"log/slog"
 )
 
 type sqlslogOptions struct {
@@ -43,50 +41,4 @@ See the following example for usage:
 */
 func Open(ctx context.Context, driverName, dsn string, opts ...Option) (*sql.DB, error) { // nolint:funlen
 	return New(driverName, dsn, opts...).Open(ctx)
-}
-
-func open(ctx context.Context, driverName, dsn string, logger *stepLogger, options *sqlslogOptions) (*sql.DB, error) {
-	lg := logger.With(
-		slog.String("driver", driverName),
-		slog.String("dsn", dsn),
-	)
-
-	var db *sql.DB
-	err := ignoreAttr(lg.Step(ctx, &options.Open, func() (*slog.Attr, error) {
-		var err error
-		db, err = openWithDriver(driverName, dsn, logger, options.DriverOptions)
-		return nil, err
-	}))
-	if err != nil {
-		return nil, err
-	}
-	return db, nil
-}
-
-func openWithDriver(driverName, dsn string, logger *stepLogger, driverOptions *driverOptions) (*sql.DB, error) {
-	db, err := sql.Open(driverName, dsn)
-	if err != nil {
-		return nil, err
-	}
-	// This db is not used directly, but it is used to get the driver.
-
-	drv := wrapDriver(db.Driver(), logger, driverOptions)
-
-	return openWithWrappedDriver(drv, dsn, logger, driverOptions)
-}
-
-func openWithWrappedDriver(drv driver.Driver, dsn string, logger *stepLogger, driverOptions *driverOptions) (*sql.DB, error) {
-	var origConnector driver.Connector
-
-	if dc, ok := drv.(driver.DriverContext); ok {
-		connector, err := dc.OpenConnector(dsn)
-		if err != nil {
-			return nil, err
-		}
-		origConnector = connector
-	} else {
-		origConnector = &dsnConnector{dsn: dsn, driver: drv}
-	}
-
-	return sql.OpenDB(wrapConnector(origConnector, logger, driverOptions.ConnectorOptions)), nil
 }
